seperate data handling from app component
